Fix #34
