+++ src/main/java/com/TheRPGAdventurer/ROTD/event/RegistryEventHandler.java
@@ -8,15 +8,12 @@ import com.TheRPGAdventurer.ROTD.inits.ModItems;
 import com.TheRPGAdventurer.ROTD.inits.ModTools;
 import com.TheRPGAdventurer.ROTD.objects.blocks.BlockDragonBreedEgg;
 import com.TheRPGAdventurer.ROTD.objects.entity.entitytameabledragon.breeds.EnumDragonBreed;
-import com.TheRPGAdventurer.ROTD.objects.items.ItemDragonAmuletNEW;
 import com.TheRPGAdventurer.ROTD.objects.items.ItemDragonBreedEgg;
 import com.TheRPGAdventurer.ROTD.objects.tileentities.TileEntityHandler;
 import com.TheRPGAdventurer.ROTD.util.DMUtils;
 import com.TheRPGAdventurer.ROTD.util.IHasModel;
 
 import net.minecraft.block.Block;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.block.model.ModelBakery;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.item.Item;
 import net.minecraft.util.ResourceLocation;
@@ -63,14 +60,10 @@ public class RegistryEventHandler {
     @SubscribeEvent
     public static void registerModels(ModelRegistryEvent event) {
         DragonMounts.proxy.registerModel(Item.getItemFromBlock(ModBlocks.DRAGONSHULKER), 0);
-        
-        // Register item render for amulet item variants
-        ModelLoader.setCustomMeshDefinition(ModItems.Amulet, new ItemDragonAmuletNEW());
-        ModelBakery.registerItemVariants(ModItems.Amulet, new ModelResourceLocation("dragonmounts:dragon_amulet"));
-        EnumDragonBreed.META_MAPPING.forEach((breed, meta) -> {
-        	ModelBakery.registerItemVariants(ModItems.Amulet, new ModelResourceLocation("dragonmounts:" + breed.getName() + "_dragon_amulet"));
-        });
 
+        // Register item render for amulet item variants
+        DragonMounts.proxy.registerAmuletRenderer();
+        
         for (Block block : ModBlocks.BLOCKS) {
         	if (block instanceof IHasModel) {
         		((IHasModel) block).RegisterModels();
