+++ src/main/java/com/TheRPGAdventurer/ROTD/objects/entity/entitytameabledragon/breeds/DragonBreedMoonlight.java
@@ -9,8 +9,6 @@ import net.minecraft.init.Blocks;
 import net.minecraft.util.EnumParticleTypes;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class DragonBreedMoonlight extends DragonBreed {
 
@@ -35,7 +33,6 @@ public class DragonBreedMoonlight extends DragonBreed {
 		if(dragon.posY > dragon.world.getHeight() + 8 && !dragon.world.isDaytime()) doParticles(dragon);
 	}
 	
-    @SideOnly(Side.CLIENT)
     private void doParticles(EntityTameableDragon dragon) {
         if (!dragon.isEgg() && !dragon.isHatchling()) {
 	        float s = dragon.getScale() * 1.2f;
